Tune COMPLEXITY terms and update a numerical inaccuracy
@TerjeKir pointed out that the combination of MG and EG in linearEvaluation() were placed into an integer instead of a double, causing a rounding of the tuning updates. This is not a major issue, but certainly hinders convergence to, and likely hides, various local and or global minima.

ELO   | 2.67 +- 2.12 (95%)
SPRT  | 12.0+0.12s Threads=1 Hash=8MB
LLR   | 2.98 (-2.94, 2.94) [0.00, 4.00]
Games | N: 35856 W: 6398 L: 6122 D: 23336
http://chess.grantnet.us/test/7071/

ELO   | 1.55 +- 1.24 (95%)
SPRT  | 60.0+0.6s Threads=1 Hash=64MB
LLR   | 2.95 (-2.94, 2.94) [0.00, 4.00]
Games | N: 75596 W: 9689 L: 9351 D: 56556
http://chess.grantnet.us/test/7072/

BENCH : 4,806,359

+++ src/tuner.c
@@ -367,7 +367,7 @@ double sigmoid(double K, double E) {
 
 double linearEvaluation(TEntry *entry, TVector params, TArray methods, TGradientData *data) {
 
-    int sign, mixed;
+    double sign, mixed;
     double midgame, endgame, wsafety[2], bsafety[2];
     double normal[PHASE_NB], safety[PHASE_NB], complexity;
     double mg[METHOD_NB][COLOUR_NB] = {0}, eg[METHOD_NB][COLOUR_NB] = {0};
@@ -391,14 +391,14 @@ double linearEvaluation(TEntry *entry, TVector params, TArray methods, TGradient
     bsafety[EG] = (double) ScoreEG(entry->safety[BLACK]) + eg[SAFETY][BLACK];
 
     // Remove the original "safety" evaluation that was double counted into the "normal" evaluation
-    normal[MG] -= MIN(0, -ScoreMG(entry->safety[WHITE]) * fabs(ScoreMG(entry->safety[WHITE])) / 720)
-                - MIN(0, -ScoreMG(entry->safety[BLACK]) * fabs(ScoreMG(entry->safety[BLACK])) / 720);
-    normal[EG] -= MIN(0, -ScoreEG(entry->safety[WHITE]) / 20) - MIN(0, -ScoreEG(entry->safety[BLACK]) / 20);
+    normal[MG] -= MIN(0, -ScoreMG(entry->safety[WHITE]) * fabs(ScoreMG(entry->safety[WHITE])) / 720.0)
+                - MIN(0, -ScoreMG(entry->safety[BLACK]) * fabs(ScoreMG(entry->safety[BLACK])) / 720.0);
+    normal[EG] -= MIN(0, -ScoreEG(entry->safety[WHITE]) / 20.0) - MIN(0, -ScoreEG(entry->safety[BLACK]) / 20.0);
 
     // Compute the new, true "safety" evaluations for each side
-    safety[MG] = MIN(0, -wsafety[MG] * fabs(-wsafety[MG]) / 720)
-               - MIN(0, -bsafety[MG] * fabs(-bsafety[MG]) / 720);
-    safety[EG] = MIN(0, -wsafety[EG] / 20) - MIN(0, -bsafety[EG] / 20);
+    safety[MG] = MIN(0, -wsafety[MG] * fabs(-wsafety[MG]) / 720.0)
+               - MIN(0, -bsafety[MG] * fabs(-bsafety[MG]) / 720.0);
+    safety[EG] = MIN(0, -wsafety[EG] / 20.0) - MIN(0, -bsafety[EG] / 20.0);
 
     // Grab the original "complexity" evaluation and add the modified parameters
     complexity = (double) ScoreEG(entry->complexity) + eg[COMPLEXITY][WHITE];
@@ -414,8 +414,8 @@ double linearEvaluation(TEntry *entry, TVector params, TArray methods, TGradient
     midgame = normal[MG] + safety[MG];
     endgame = normal[EG] + safety[EG] + sign * fmax(-fabs(normal[EG] + safety[EG]), complexity);
 
-    mixed = (midgame * (256 - entry->phase)
-          +  endgame * entry->phase * entry->sfactor) / 256;
+    mixed = (midgame * (256.0 - entry->phase)
+          +  endgame * entry->phase * entry->sfactor) / 256.0;
 
     return mixed + (entry->turn == WHITE ? Tempo : -Tempo);
 }
